% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_sigma.R
\name{gen_sigma}
\alias{gen_sigma}
\title{Generate covariance matrix for observed treatment effect in phase 3}
\usage{
gen_sigma(n_events, ratio = 1)
}
\arguments{
\item{n_events}{numeric vector of target number of events in phase 3}

\item{ratio}{randomization ratio of experimental arm compared to control}
}
\value{
sigma_unit and covariance matrix to be used for PoS prediction
}
\description{
Generates covariance matrix for an observed treatment effect,
i.e., log of hazard ratio in a phase 3 trial
}
\section{Specification}{

\if{latex}{
  \itemize{
  \item Transfer the randomization ratio to the proportion of subjects in 
  the control group
  \item Calculate the unit standard deviation based on the proportion of 
  subjects in the control group
  \item Create a covariance matrix for the observed treatment effect (log 
  hazard ratio)
  \item Update the matrix elements based on the target number of events in a
 phase 3 study
  }
}
\if{html}{
The contents of this section are shown in PDF user manual only.
}
}

\examples{
\dontrun{gen_sigma(c(100, 150), ratio = 1)}
}
