% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_denominator.R
\name{get_denominator}
\alias{get_denominator}
\title{Transform degree of heterogeneity to denominator}
\usage{
get_denominator(input)
}
\arguments{
\item{input}{indicates the degree of heterogeneity of the study level
parameter, and must be one of "large", "substantial", "moderate", "small",
"very small"}
}
\description{
Helper function: transform degree of heterogeneity to
denominator
}
\section{Specification}{

\if{latex}{
\itemize{
\item Determine the denominator of standard deviation for tau (standard 
deviation of phase 2 or 3 treatment effect) based on the level of 
heterogeneity 
}
}
\if{html}{
The contents of this section are shown in PDF user manual only.
}
}

