% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resp2oddsratio.R
\name{resp2oddsratio}
\alias{resp2oddsratio}
\title{Estimate log odds ratio given response}
\usage{
resp2oddsratio(n_resp_trt, n_resp_ctrl, n_trt, n_ctrl)
}
\arguments{
\item{n_resp_trt}{number of responses in treatment arm}

\item{n_resp_ctrl}{number of responses in control arm}

\item{n_trt}{sample size in treatment arm}

\item{n_ctrl}{sample size in control arm}
}
\value{
log odds ratio point estimate and a corresponding standard error
}
\description{
Estimates log odds ratio for treatment vs control based on
number of responses and samples size in each arm
}
\section{Specification}{

\if{latex}{
  \itemize{
  \item Calculate proportion of responders in treatment and control groups
 respectively
  \item Derive ORR odds in each arm
  \item Calculate observed ORR odds ratio and the corresponding standard error 
  }
}
\if{html}{
The contents of this section are shown in PDF user manual only.
}
}

\examples{
resp2oddsratio(n_resp_trt = 20, n_resp_ctrl = 10, n_trt = 40, n_ctrl = 45)
}
