% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_pos.R
\name{gen_pos}
\alias{gen_pos}
\title{PoS estimation}
\usage{
gen_pos(
  target_hr,
  J,
  nevents3,
  hr_bound,
  omega,
  est_obs_pfs,
  low_obs_pfs,
  upp_obs_pfs,
  obs_pfs_conf_level = 0.95,
  thres = 0.01,
  n_trt2,
  n_ctrl2,
  n_resp_trt2,
  n_resp_ctrl2,
  use_orr = FALSE,
  use_pfs = FALSE,
  het_degree_p2 = "small",
  het_degree_p3 = "very small",
  ratio = 1,
  m_0 = NA,
  m_1 = NA,
  nu_0 = NA,
  nu_1 = NA,
  lm_sd = NA,
  niter = 1000,
  nchains = 4,
  ncores = 4,
  seed,
  plots_out = FALSE,
  ...
)
}
\arguments{
\item{target_hr}{target hazard ration in phase 3 study (i.e., under
alternative hypothesis)}

\item{J}{number of planned analyses in phase 3 study}

\item{nevents3}{numeric vector of target number of events in phase 3 study}

\item{hr_bound}{numeric vector of hazard ratio bounds for analyses in phase
3 study}

\item{omega}{probability that a treatment effect comes from a enthusiastic
prior component, i.e., initial benchmarking probability for the study success}

\item{est_obs_pfs}{estimated PFS hazard ratio based on prior/earlier study}

\item{low_obs_pfs}{lower bound of estimated PFS hazard ratio based on
prior/earlier study}

\item{upp_obs_pfs}{upper bound of estimated PFS hazard ratio based on
prior/earlier study}

\item{obs_pfs_conf_level}{confidence level for the estimated PFS hazard ratio
bounds, Default: 0.95}

\item{thres}{probability of either lack treatment effect under the
enthusiastic prior or substantial treatment effect under the pessimistic
prior. Should be set to a small value, close to 0, Default: 0.01}

\item{n_trt2}{sample size in treatment arm from a prior/earlier study}

\item{n_ctrl2}{sample size in control arm from a prior/earlier study}

\item{n_resp_trt2}{number of responses in treatment arm from a prior/earlier
study}

\item{n_resp_ctrl2}{number of responses in control arm from a prior/earlier
study}

\item{use_orr}{whether response data from a prior/earlier study should be
used, Default: FALSE}

\item{use_pfs}{whether PFS data from a prior/earlier study should be used,
Default: FALSE}

\item{het_degree_p2}{indicates the degree of heterogeneity of a study level
parameter for a prior/earlier study and must be one of "large",
"substantial", "moderate", "small", "very small", Default: 'small'.}

\item{het_degree_p3}{indicates the degree of heterogeneity of a study level
parameter for a phase 3 study and must be one of "large",
"substantial", "moderate", "small", "very small", Default: 'very small'}

\item{ratio}{randomization ratio of experimental arm compared to control}

\item{m_0}{intercept for linear regression of log treatment effect of PFS on
log treatment effect on response. A value is expected only when \code{use_orr = TRUE}
and \code{use_pfs = TRUE}, Default: NA}

\item{m_1}{slope for linear regression of log treatment effect of PFS on
log treatment effect on response. A value is expected only when \code{use_orr = TRUE}
and \code{use_pfs = TRUE}, Default: NA}

\item{nu_0}{standard error of \code{m_0}. A value is expected only when
\code{use_orr = TRUE} and \code{use_pfs = TRUE}, Default: NA}

\item{nu_1}{standard error of \code{m_1}. A value is expected only when
\code{use_orr = TRUE} and \code{use_pfs = TRUE}, Default: NA}

\item{lm_sd}{linear regression residual variance of log treatment effect of
PFS on log treatment effect on response. A value is expected only when
\code{use_orr = TRUE} and \code{use_pfs = TRUE}, Default: NA}

\item{niter}{number of iterations to be used in stan run, Default: 1000}

\item{nchains}{number of chains to be used in stan run, Default: 4}

\item{ncores}{number of cores to be used in stan run, Default: 4}

\item{seed}{seed to be used in stan run}

\item{plots_out}{whether plots for MCMC chains and autocorrelation should be
printed, Default: FALSE}

\item{...}{params to pass to stan run}
}
\value{
tibble of PoS estimates and the corresponding standard errors for
each analysis. If \code{plots_out} is turned on, then the MCMC chains mixing and
autocorrelation plots are provided as well.
}
\description{
PoS estimation
}
\section{Specification}{

\if{latex}{
  \itemize{
  \item Obtain the prediction of PoS using `run_stan()`
  \item Create an object `fit_stan` to save the model result
  \item Transfer the hazard ratio bound to log-scale
  \item Extract simulated phase 3 trials treatment effects estimates and
  derive their standard errors
  \item If `plots_out` is TRUE, generate plots for MCMC chains and 
  autocorrelation
  }
}
\if{html}{
The contents of this section are shown in PDF user manual only.
}
}

\examples{
# use PFS data from a prior study

  gen_pos(
    target_hr = 0.7, 
    J = 2,  
    nevents3 = c(370, 468), 
    hr_bound = c(0.779, 0.8204), 
    omega = 0.5, 
    est_obs_pfs = 0.88, 
    low_obs_pfs = 0.74, 
    upp_obs_pfs = 1.05, 
    use_pfs = TRUE, 
    seed = 325,
    ncores = 1,
    nchains = 1) 
}
\seealso{
\code{\link[tidybayes]{gather_draws}}
\code{\link[dplyr]{mutate}}, \code{\link[dplyr]{group_by}}, \code{\link[dplyr]{summarise}}
\code{\link[purrr]{map2}}, \code{\link[purrr]{reexports}}
\code{\link[bayesplot]{MCMC-traces}}, \code{\link[bayesplot]{MCMC-diagnostics}}
}
