% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_stan.R
\name{run_stan}
\alias{run_stan}
\title{Run Stan to Predict Probability of Success (PoS) for Phase III Trial}
\usage{
run_stan(
  target_hr,
  J,
  nevents3,
  hr_bound,
  omega_mean = 0.52,
  omega_var = 0.02,
  est_obs_pfs,
  low_obs_pfs,
  upp_obs_pfs,
  obs_pfs_conf_level = 0.95,
  thres = 0.01,
  n_trt2,
  n_ctrl2,
  n_resp_trt2,
  n_resp_ctrl2,
  low_soc_rr,
  upp_soc_rr,
  ci_rr = 0.8,
  use_orr = FALSE,
  single_arm = FALSE,
  use_pfs = FALSE,
  het_degree_p2 = "small",
  het_degree_p3 = "very small",
  ratio = 1,
  indication = 6,
  m_0 = NA,
  m_1 = NA,
  nu_0 = NA,
  nu_1 = NA,
  lm_sd = NA,
  niter = 1000,
  nchains = 4,
  ncores = 4,
  seed,
  ...
)
}
\arguments{
\item{target_hr}{target hazard ration in phase 3 study (i.e., under
alternative hypothesis)}

\item{J}{number of planned analyses in phase 3 study}

\item{nevents3}{numeric vector of target number of events in phase 3 study}

\item{hr_bound}{numeric vector of hazard ratio bounds for analyses in phase
3 study}

\item{omega_mean}{Mean of the Beta prior for \code{omega}. \code{omega} is
the probability that a treatment effect comes from a enthusiastic prior
component, i.e., initial benchmarking probability for the study success}

\item{omega_var}{Variance of the Beta prior for \code{omega}.}

\item{est_obs_pfs}{estimated PFS hazard ratio based on prior/earlier study}

\item{low_obs_pfs}{lower bound of estimated PFS hazard ratio based on
prior/earlier study}

\item{upp_obs_pfs}{upper bound of estimated PFS hazard ratio based on
prior/earlier study}

\item{obs_pfs_conf_level}{confidence level for the estimated PFS hazard ratio
bounds, Default: 0.95}

\item{thres}{probability of either lack treatment effect under the
enthusiastic prior or substantial treatment effect under the pessimistic
prior. Should be set to a small value, close to 0, Default: 0.01}

\item{n_trt2}{sample size in treatment arm from a prior/earlier study}

\item{n_ctrl2}{sample size in control arm from a prior/earlier study}

\item{n_resp_trt2}{number of responses in treatment arm from a prior/earlier
study}

\item{n_resp_ctrl2}{number of responses in control arm from a prior/earlier
study}

\item{low_soc_rr}{Lower bound of historical control response rate for
single-arm estimation.}

\item{upp_soc_rr}{Upper bound of historical control response rate for
single-arm estimation.}

\item{ci_rr}{Confidence level for control response rate bounds, Default: 0.8.}

\item{use_orr}{whether response data from a prior/earlier study should be
used, Default: FALSE}

\item{single_arm}{whether ORR data is from a single-arm trial, Default:
FALSE.}

\item{use_pfs}{whether PFS data from a prior/earlier study should be used,
Default: FALSE}

\item{het_degree_p2}{indicates the degree of heterogeneity of a study level
parameter for a prior/earlier study and must be one of "large",
"substantial", "moderate", "small", "very small", Default: 'small'.}

\item{het_degree_p3}{indicates the degree of heterogeneity of a study level
parameter for a phase 3 study and must be one of "large",
"substantial", "moderate", "small", "very small", Default: 'very small'}

\item{ratio}{randomization ratio of experimental arm compared to control}

\item{indication}{Integer from 1 to 6 for selecting indication-specific
ORR–PFS regression parameters, as follows:
1 = hematologic malignancies
2 = gynecologic cancers
3 = thoracic cancers
4 = other solid tumors
5 = breast cancer
6 = any tumor type
Default: 6}

\item{m_0}{intercept for linear regression of log treatment effect of PFS on
log treatment effect on response. A value is expected only when
\code{use_orr = TRUE} and \code{use_pfs = TRUE}. Auto-filled based on \code{indication}
if not supplied.}

\item{m_1}{slope for linear regression of log treatment effect of PFS on
log treatment effect on response. A value is expected only when
\code{use_orr = TRUE} and \code{use_pfs = TRUE}. Auto-filled based on \code{indication}
if not supplied.}

\item{nu_0}{standard error of \code{m_0}. A value is expected only when
\code{use_orr = TRUE} and \code{use_pfs = TRUE}. Auto-filled based on \code{indication}
if not supplied.}

\item{nu_1}{standard error of \code{m_1}. A value is expected only when
\code{use_orr = TRUE} and \code{use_pfs = TRUE}. Auto-filled based on \code{indication}
if not supplied.}

\item{lm_sd}{linear regression residual variance of log treatment effect of
PFS on log treatment effect on response. A value is expected only when
\code{use_orr = TRUE} and \code{use_pfs = TRUE}. Auto-filled based on \code{indication}
if not supplied.}

\item{niter}{number of iterations to be used in stan run, Default: 1000}

\item{nchains}{number of chains to be used in stan run, Default: 4}

\item{ncores}{number of cores to be used in stan run, Default: 4}

\item{seed}{seed to be used in stan run}

\item{...}{params to pass to stan run}
}
\value{
list which includes the generated stan object, list of data that
was supplied to \code{rstan::stan()}, and the name of the stan file which was run
}
\description{
A wrapper function to run stan in order to generate a PoS
prediction
}
\section{Specification}{

\if{latex}{
  \itemize{
  \item Transfer target hazard ratio and hazard ratio bound to log-scale
  \item When `use_PFS` is TRUE, calculate log hazard ratio PFS estimate and  
  its standard error from a phase 2 study
  \item When `use_ORR` is TRUE, calculate log odds ratio for ORR estimate 
  and its standard error estimate using either two-arm or single-arm data
  from a phase 2 study
  \item Construct a Beta prior for \code{omega} using user-specified
  mean and variance.
  \item Derive sigma_P1 and sigma_P2, i.e., the standard deviations in the 
  mixture prior, based on the input threshold value
  \item Calculate the unit standard deviation and the covariance matrix of 
  the observed treatment effect using `gen_sigma()`
  \item Select indication-specific regression parameters for surrogate
  -primary endpoint relationship when the primary endpoint (e.g., PFS) is
  not directly available from the earlier-phase study.
  \item Derive the hyperparameter for the half normal distribution
  \item Create a list for `rstan::stan()` input
  \item Update the list based on `use_PFS` and `use_ORR`
  \item Run `rstan::stan()` to obtain the PoS prediction
  }
}
\if{html}{
The contents of this section are shown in PDF user manual only.
}
}

\examples{
# use single arm ORR data from a prior study
  run_stan(
  target_hr =  0.70,
  J = 2,
  nevents3 = c(370, 468),
  hr_bound = c(0.7790, 0.8204),
  thres = 0.01, 
  n_trt2 = 100, 
  n_resp_trt2 = 40, 
  low_soc_rr = 0.05, 
  upp_soc_rr = 0.2, 
  use_orr = TRUE, 
  single_arm = TRUE,
  seed=111)
}
\seealso{
\code{\link[stats]{Normal}}
\code{\link[pracma]{erf}}
\code{\link[rstan]{stan}}
}
